# src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: Customer Booking System API
  description: API for managing customer reservations in the booking system.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api

paths:
  /reservations:
    post:
      summary: Create a new reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: John Doe
                phoneNumber:
                  type: string
                  example: 1234567890
                email:
                  type: string
                  example: john.doe@example.com
                reservationDateTime:
                  type: string
                  format: date-time
                  example: 2024-10-21T19:30:00
                numberOfGuests:
                  type: integer
                  example: 4
      responses:
        '200':
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '400':
          description: Invalid request
        '500':
          description: Internal server error
    get:
      summary: Get all reservations by email
      parameters:
        - name: email
          in: query
          required: true
          description: The email of the customer to filter reservations.
          schema:
            type: string
      responses:
        '200':
          description: A list of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reservation'

  /reservations/{id}:
    delete:
      summary: Cancel a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Reservation cancelled
        '404':
          description: Reservation not found
    put:
      summary: Update a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reservationDateTime:
                  type: string
                  format: date-time
                  example: 2024-10-21T21:30:00
                numberOfGuests:
                  type: integer
                  example: 3
      responses:
        '200':
          description: Reservation updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '404':
          description: Reservation not found

components:
  schemas:
    Reservation:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: John Doe
        phoneNumber:
          type: string
          example: 1234567890
        email:
          type: string
          example: john.doe@example.com
        reservationDateTime:
          type: string
          format: date-time
          example: 2024-10-21T19:30:00
        numberOfGuests:
          type: integer
          example: 4
        status:
          type: string
          example: CONFIRMED